Remove ChecksummedSlice
