+++ mountpoint-s3/tests/reftests/harness.rs
@@ -6,7 +6,6 @@ use std::sync::Arc;
 use fuser::FileType;
 use futures::executor::ThreadPool;
 use futures::future::{BoxFuture, FutureExt};
-use mountpoint_s3::checksums::ChecksummedSlice;
 use mountpoint_s3::fs::{InodeNo, ReadReplier, FUSE_ROOT_INODE};
 use mountpoint_s3::prefix::Prefix;
 use mountpoint_s3::{S3Filesystem, S3FilesystemConfig};
@@ -343,7 +342,7 @@ impl Harness {
                 inflight_write.inode,
                 file_handle,
                 inflight_write.written as i64,
-                ChecksummedSlice::new(&bytes_to_write),
+                &bytes_to_write,
                 0,
                 0,
                 None,
